1.6-text_to_image/pyproject.toml
[project]
name = "image-generator"
version = "0.1.0"
description = "Text-to-Image Generator using Hugging Face Inference API"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "python-dotenv==1.0.0",
    "huggingface-hub>=0.20.0",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "mypy>=1.19.0",
    "ruff>=0.14.8"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py310"
